refactor the service layer

+++ src/main/java/vn/unigap/api/dto/in/SeekerDtoIn.java
@@ -28,5 +28,5 @@ public class SeekerDtoIn {
     private String address;
 
     @NotNull
-    private Integer provinceId;
+    private Long provinceId;
 }
